refactor(lib): extract GitHub URL helpers into url module (#217)
Move github_owner_repo / looks_like_local_path / is_valid_github_owner /
is_valid_github_repo / installed_full_name / urls_match / format_plugin_url
out of lib.rs into src/url.rs (all pub(crate), glob-imported by lib.rs).

Behaviour-preserving code motion. fmt --check + clippy -D warnings +
test (805) + --doc all green.
